% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crible.R
\name{CA}
\alias{CA}
\title{Crible d'Atkin
Génère les nombres premiers inférieurs ou égal a un entier positif donné nmax}
\usage{
CA(nmax)
}
\arguments{
\item{nmax}{}
}
\value{
une seule valeur: juste la valeur d'entrée
}
\description{
Crible d'Atkin
Génère les nombres premiers inférieurs ou égal a un entier positif donné nmax
}
\author{
NDJEBAYI PATRICK NATANAEL
}
