% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaPa-package.R
\docType{package}
\name{NaPa-package}
\alias{NaPa}
\alias{NaPa-package}
\title{NaPa: The package encrypts and decrypts messages}
\description{
the package allows for encrypting and decrypting messages using specific fuctions such as CA(),which is an Atkin sieve function that generate prime numbers lower than a given number.
}
\keyword{internal}
